Объект двигается плавно и медленно

+++ app/src/main/java/com/itsamsung/robot/Mech.java
@@ -4,7 +4,7 @@ package com.itsamsung.robot;
  * Created by TEACHER on 30.11.2016.
  */
 public abstract class Mech {
-    protected int size;
-    protected int x;
-    protected int y;
+    protected float size;
+    protected float x;
+    protected float y;
 }
